[Update] move hook folder to component folder
